+++ developing/frontend/kitoAi/src/app/app.component.ts
@@ -7,6 +7,7 @@ import {MatSidenavModule} from "@angular/material/sidenav";
 import {MatListModule} from "@angular/material/list";
 import {MatButtonModule} from "@angular/material/button";
 import {AuthenticationService} from "./core/auth/authentication.service";
+import {MENU, MenuItem} from "./shared/menu";
 
 @Component({
   selector: 'kito-root',
@@ -24,25 +25,30 @@ import {AuthenticationService} from "./core/auth/authentication.service";
 	],
   template: `
       <div class="wrapper">
-          <mat-toolbar color="primary">
+          <mat-toolbar color="primary" class="toolbar">
 
-              <button mat-icon-button (click)="sidenav.toggle()">
-                  <mat-icon>menu</mat-icon>
-              </button>
+              <div class="toolbar-left">
+                  <button mat-icon-button (click)="sidenav.toggle()">
+                      <mat-icon>menu</mat-icon>
+                  </button>
 
-              <h2>{{title}}</h2>
-              <mat-icon>donut_small</mat-icon>
+                  <h2>{{title}}</h2>
+                  <mat-icon>donut_small</mat-icon>
+              </div>
 
-              <span class="flex"></span>
-              <button mat-icon-button (click)="sidenav.toggle()">
-                  <mat-icon>user</mat-icon>
-              </button>
+              <div class="toolbar-right">
+                  <button mat-icon-button
+                          *ngIf="(authServices.isLogged$() | async)"
+                          (click)="authServices.logout()">
+                      <mat-icon>logout</mat-icon>
+                  </button>
+              </div>
 
           </mat-toolbar>
 
           <div class="container">
 			  <mat-sidenav-container class="sidenav-container">
-				  <mat-sidenav #sidenav mode="push">
+				  <mat-sidenav #sidenav mode="over" class="sidenav">
 					  <mat-nav-list>
 						  <a mat-list-item routerLink="/public">Public</a>
 						  <mat-divider></mat-divider>
@@ -51,8 +57,9 @@ import {AuthenticationService} from "./core/auth/authentication.service";
 								 routerLink="/auth/login">Login</a>
 						  </ng-container>
 						  <ng-template #logged>
-							  <a mat-list-item routerLink="/">Home</a>
-							  <a mat-list-item routerLink="/projects">Projects</a>
+                              <a *ngFor="let item of menuItems" mat-list-item [routerLink]="item.link">
+                                  {{item.label}}
+                              </a>
 							  <mat-divider></mat-divider>
 							  <a mat-list-item (click)="authServices.logout()">LogOut</a>
 						  </ng-template>
@@ -60,30 +67,54 @@ import {AuthenticationService} from "./core/auth/authentication.service";
 				  </mat-sidenav>
 				  <!-- Main Content Area -->
 				  <mat-sidenav-content>
-
-					  <router-outlet></router-outlet>
-
+					  <router-outlet class="router"></router-outlet>
 				  </mat-sidenav-content>
 			  </mat-sidenav-container>
           </div>
 
           <mat-toolbar color="primary" class="footer">
-              <!-- Footer Content -->
-              <p class="text-center text-md-left">Copyright © {{date.getFullYear()}}kito.ai. All rights reserved</p>
-              <p class="text-center center">v0.0.2</p>
-              <p class="text-center center">Build #: 1</p>
-              <p class="text-center center">Build: 11-16-2023</p>
-              <p class="text-center text-md-left mb-0 ">Handcrafted With <i
-                      class="mb-1 text-primary ml-1 icon-small feather icon-heart"></i></p>
+              <p class="footer-left">Copyright © {{date.getFullYear()}}kito.ai. All rights reserved</p>
+
+              <div class="footer-center">
+                  <p class="mt-2">v0.0.2</p>
+                  <p class="mt-2">Build #: 1</p>
+                  <p class="mt-2">Build: 11-16-2023</p>
+              </div>
+
           </mat-toolbar>
+
+
       </div>
   `,  styles: `
-
 	.wrapper {
 	  display: flex;
 	  flex-direction: column;
 	  height: 100vh;
 	}
+
+
+  	.router{
+	  padding: 10px;
+  	}
+
+	.toolbar {
+	  position: fixed;
+	  top: 0;
+	  width: 100%;
+	  height: 50px;
+	  color: white;
+	  z-index: 10;
+	}
+
+	.toolbar-left {
+      display: flex; /* Create left section */
+      align-items: center;
+    }
+
+    .toolbar-right {
+      margin-left: auto;
+      align-items: right;
+    }
 
 	.container {
 	  flex: 1;
@@ -92,25 +123,50 @@ import {AuthenticationService} from "./core/auth/authentication.service";
 
 	.sidenav-container {
 	  flex: 1;
+	  top: 50px;
 	}
 
-	.footer {
+	.sidenav{
+	    top: 50px;
+	    position: fixed;
+	}
+
+    .footer {
+	  position: fixed;
+	  bottom: 0;
 	  width: 100%;
-	  margin-top: auto;
+	  height: 30px;
+	  font-size: 12px;
+	  color: lightgray;
+	  z-index: 9;
 	}
 
-	.spacer {
-	  flex: 1 1 auto;
-	}
+	.footer-left {
+      display: flex; /* Create first column */
+      text-align: left;
+    }
+
+    .footer-center {
+      display: flex; /* Create second column */
+      text-align: center;
+    }
+
+    .footer-right {
+      display: flex; /* Create third column */
+      text-align: right;
+    }
+
   `,
 })
 export class AppComponent implements  OnInit{
-  title = 'kitoAi';
+    title = 'kitoAi';
 
-  date: Date = new Date()
+    date: Date = new Date()
 
-  constructor(public authServices: AuthenticationService) {}
+    menuItems: MenuItem[] = MENU.map(x => x);
 
-	ngOnInit() {
-	}
+    constructor(public authServices: AuthenticationService) {}
+
+    ngOnInit() {
+    }
 }