+++ developing/frontend/kitoAi/src/app/app.config.ts
@@ -20,7 +20,7 @@ export const appConfig: ApplicationConfig = {
 	  provideHttpClient(withInterceptors([
 		  TokenInterceptor,
 		  ErrorInterceptor,
-	  ]))
+	  ]), withFetch())
 	  //provideClientHydration(),
 
   ]