+++ developing/frontend/kitoAi/src/app/app.routes.ts
@@ -1,22 +1,25 @@
 import { Routes } from '@angular/router';
 import {AuthGuard} from "./core/guard/authGuard";
-import {BaseComponent} from "./features/layout/base.component";
 import {ErrorPageComponent} from "./features/error/error-page.component";
 import {LoginComponent} from "./features/users/components/login.component";
-import {AppComponent} from "./app.component";
 import {HomeComponent} from "./features/home/components/home.component";
+import {DashboardComponent} from "./features/home/components/dashboard.component";
 import {ProjectsComponent} from "./features/projects/components/projects.component";
-import {PROJECTS_ROUTES} from "./features/projects/projects.routes";
-
-const projectsRoutes = () => import('./features/projects/projects.routes').then(x => x.PROJECTS_ROUTES);
-const publicRoutes = () => import('./features/public/public.routes').then(x => x.PUBLIC_ROUTES);
+import {PublicComponent} from "./features/public/components/public.component";
 
 
 export const APP_ROUTES: Routes = [
+	// AUTH routes
 	{ path: 'auth/login', component: LoginComponent },
-	{ path: 'home', canActivate: [AuthGuard], component: HomeComponent },
-	{ path: 'project', canActivate: [AuthGuard], loadChildren: projectsRoutes },
-	{ path: 'public', loadChildren: publicRoutes },
+	// PUBLIC routes
+	{ path: 'public', component: PublicComponent},
+	// ADMIN routes
+    // home
+	{ path: 'home', canActivate: [AuthGuard], component: HomeComponent},
+    { path: 'home/dashboard', canActivate: [AuthGuard], component: DashboardComponent},
+    // projects
+	{ path: 'projects', canActivate: [AuthGuard], component: ProjectsComponent},
+	// UTIL routes
 	{
 		path: 'error',
 		component: ErrorPageComponent,