+++ developing/frontend/kitoAi/src/app/features/home/components/dashboard.component.ts
@@ -2,16 +2,16 @@ import { Component } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 @Component({
-  selector: 'kito-base',
+  selector: 'kito-dashboard',
   standalone: true,
   imports: [CommonModule],
   template: `
     <p>
-      base works!
+      dashboard works!
     </p>
   `,
   styles: ``
 })
-export class BaseComponent {
+export class DashboardComponent {
 
 }