+++ developing/frontend/kitoAi/src/app/features/public/public.routes.ts
@@ -1,7 +1,0 @@
-import {Routes} from "@angular/router";
-import {PublicComponent} from "./components/public.component";
-
-export const PUBLIC_ROUTES: Routes = [
-	{ path: '', component: PublicComponent},
-	{ path: '**', redirectTo: '' }
-];